+++ src/modules/config/config.controller.ts
@@ -9,11 +9,17 @@ import * as configService from './config.service'
 import { IConfigDoc } from './config.interfaces';
 
 export const createConfig = catchAsync(async (req: Request, res: Response<IConfigDoc>) => {
+    if (req.scope == 'company' || req.scope == 'building') {
+        throw new ApiError(httpStatus.FORBIDDEN, 'Cannot create config')
+    }
     const config = await configService.createConfig(req.body);
     res.status(httpStatus.CREATED).send(config);
 });
 
 export const getConfigs = catchAsync(async (req: Request, res: Response) => {
+    if (req.scope == 'company' || req.scope == 'building') {
+        throw new ApiError(httpStatus.FORBIDDEN, 'Cannot get configs')
+    }
     const filter = pick(req.query, ['key']);
     const options: IOptions = pick(req.query, ['sortBy', 'limit', 'page', 'projectBy']);
     const result = await configService.queryConfigs(filter, options);
@@ -22,6 +28,9 @@ export const getConfigs = catchAsync(async (req: Request, res: Response) => {
 
 export const getConfigById = catchAsync(async (req: Request, res: Response<IConfigDoc>) => {
     if (typeof req.params['configId'] === 'string') {
+        if (req.scope == 'company' || req.scope == 'building') {
+            throw new ApiError(httpStatus.FORBIDDEN, 'Cannot get config')
+        }
         const config = await configService.getConfigById(new mongoose.Types.ObjectId(req.params['configId']));
         if (!config) {
             throw new ApiError(httpStatus.NOT_FOUND, 'Config not found');
@@ -32,6 +41,9 @@ export const getConfigById = catchAsync(async (req: Request, res: Response<IConf
 
 export const updateConfig = catchAsync(async (req: Request, res: Response<IConfigDoc | null>) => {
     if (typeof req.params['configId'] === 'string') {
+        if (req.scope == 'company' || req.scope == 'building') {
+            throw new ApiError(httpStatus.FORBIDDEN, 'Cannot update config')
+        }
         const config = await configService.updateConfigById(new mongoose.Types.ObjectId(req.params['configId']), req.body);
         res.send(config);
     }
@@ -39,6 +51,9 @@ export const updateConfig = catchAsync(async (req: Request, res: Response<IConfi
 
 export const deleteConfig = catchAsync(async (req: Request, res: Response<null>) => {
     if (typeof req.params['configId'] === 'string') {
+        if (req.scope == 'company' || req.scope == 'building') {
+            throw new ApiError(httpStatus.FORBIDDEN, 'Cannot delete config')
+        }
         await configService.deleteConfigById(new mongoose.Types.ObjectId(req.params['configId']));
         res.status(httpStatus.NO_CONTENT).send();
     }
